// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/owner-reconciliation-status.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AccessOwnerReconciliationSummary } from '../models/access-owner-reconciliation-summary';
import { OwnerReconciliationStatusCandidate } from '../models/owner-reconciliation-status-candidate';
export interface OwnerReconciliationStatus {
  canApply: boolean;
  candidates: Array<OwnerReconciliationStatusCandidate>;
  configurationFingerprint: string;
  isReady: boolean;
  lastReconciliation: AccessOwnerReconciliationSummary | null;
}